pyproject.toml
[tool.poetry]
name = "repl_nix_CivBot"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"
disnake = "^2.7.0"
plyvel = "^1.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
